requirements.txt
PyNaCl==1.5.0
requests==2.22.0
pygsheets==2.0.5
python-telegram-bot==20.0a2

